moved fetching to chatbot-dialog
moved the fetching functionality from the modeler to the chatbot-diaglog file
